refactor insert char. This needs to be commented
